+++ synapse-redis/src/main/java/de/otto/synapse/messagestore/redis/RedisRingBufferMessageStore.java
@@ -167,8 +167,8 @@ public class RedisRingBufferMessageStore implements MessageStore {
     }
 
     @Override
-    public int size() {
-        return redisTemplate.boundListOps(name + "-messages").size().intValue();
+    public long size() {
+        return redisTemplate.boundListOps(name + "-messages").size();
     }
 
     @Override
